Refactor code and add FastAPI as API framework
Refactored the existing 'ouo-bypass.py' script by renaming it to 'bypass.py' for better understanding and commented out the lines that were outputting result to console as this is now handled by FastAPI. Then, fastAPI implementation was added in 'main.py' to enable the bypass functionality to be used as a REST service. The 'requirements.txt' file was then updated to reflect the libraries needed for the FastAPI and pydantic.